+++ tools/downloader/src/open_model_zoo/model_tools/downloader.py
@@ -16,272 +16,16 @@
 
 import argparse
 import contextlib
-import functools
-import hashlib
 import json
-import requests
-import shutil
-import ssl
 import sys
-import tempfile
-import threading
-import time
-import types
 
 from pathlib import Path
 
 from open_model_zoo.model_tools import (
     _configuration, _common, _concurrency, _reporting,
 )
-
-
-CHUNK_SIZE = 1 << 15 if sys.stdout.isatty() else 1 << 20
-
-def process_download(reporter, chunk_iterable, size, progress, file):
-    start_time = time.monotonic()
-    start_size = progress.size
-
-    try:
-        for chunk in chunk_iterable:
-            reporter.job_context.check_interrupted()
-
-            if chunk:
-                duration = time.monotonic() - start_time
-                progress.size += len(chunk)
-                progress.hasher.update(chunk)
-
-                if duration != 0:
-                    speed = int((progress.size - start_size) / (1024 * duration))
-                else:
-                    speed = '?'
-
-                percent = progress.size * 100 // size
-
-                reporter.print_progress('... {}%, {} KB, {} KB/s, {} seconds passed',
-                    percent, progress.size // 1024, speed, int(duration))
-                reporter.emit_event('model_file_download_progress', size=progress.size)
-
-                file.write(chunk)
-
-                # don't attempt to finish a file if it's bigger than expected
-                if progress.size > size:
-                    break
-    finally:
-        reporter.end_progress()
-
-def try_download(reporter, file, num_attempts, start_download, size):
-    progress = types.SimpleNamespace(size=0)
-
-    for attempt in range(num_attempts):
-        if attempt != 0:
-            retry_delay = 10
-            reporter.print("Will retry in {} seconds...", retry_delay, flush=True)
-            time.sleep(retry_delay)
-
-        try:
-            reporter.job_context.check_interrupted()
-            chunk_iterable, continue_offset = start_download(offset=progress.size)
-
-            if continue_offset not in {0, progress.size}:
-                # Somehow we neither restarted nor continued from where we left off.
-                # Try to restart.
-                chunk_iterable, continue_offset = start_download(offset=0)
-                if continue_offset != 0:
-                    reporter.log_error("Remote server refuses to send whole file, aborting")
-                    return None
-
-            if continue_offset == 0:
-                file.seek(0)
-                file.truncate()
-                progress.size = 0
-                progress.hasher = hashlib.sha256()
-
-            process_download(reporter, chunk_iterable, size, progress, file)
-
-            if progress.size > size:
-                reporter.log_error("Remote file is longer than expected ({} B), download aborted", size)
-                # no sense in retrying - if the file is longer, there's no way it'll fix itself
-                return None
-            elif progress.size < size:
-                reporter.log_error("Downloaded file is shorter ({} B) than expected ({} B)",
-                    progress.size, size)
-                # it's possible that we got disconnected before receiving the full file,
-                # so try again
-            else:
-                return progress.hasher.digest()
-        except (requests.exceptions.RequestException, ssl.SSLError):
-            reporter.log_error("Download failed", exc_info=True)
-
-    return None
-
-def verify_hash(reporter, actual_hash, expected_hash, path):
-    if actual_hash != expected_hash:
-        reporter.log_error('Hash mismatch for "{}"', path)
-        reporter.log_details('Expected: {}', expected_hash.hex())
-        reporter.log_details('Actual:   {}', actual_hash.hex())
-        return False
-    return True
-
-class NullCache:
-    def has(self, hash): return False
-    def get(self, model_file, path, reporter): return False
-    def put(self, hash, path): pass
-
-class DirCache:
-    _FORMAT = 1 # increment if backwards-incompatible changes to the format are made
-    _HASH_LEN = hashlib.sha256().digest_size
-
-    def __init__(self, cache_dir):
-        self._cache_dir = cache_dir / str(self._FORMAT)
-        self._cache_dir.mkdir(parents=True, exist_ok=True)
-
-        self._staging_dir = self._cache_dir / 'staging'
-        self._staging_dir.mkdir(exist_ok=True)
-
-    def _hash_path(self, hash):
-        assert len(hash) == self._HASH_LEN
-        hash_str = hash.hex().lower()
-        return self._cache_dir / hash_str[:2] / hash_str[2:]
-
-    def has(self, hash):
-        return self._hash_path(hash).exists()
-
-    def get(self, model_file, path, reporter):
-        cache_path = self._hash_path(model_file.sha256)
-        cache_sha256 = hashlib.sha256()
-        cache_size = 0
-
-        with open(cache_path, 'rb') as cache_file, open(path, 'wb') as destination_file:
-            while True:
-                data = cache_file.read(CHUNK_SIZE)
-                if not data:
-                    break
-                cache_size += len(data)
-                if cache_size > model_file.size:
-                    reporter.log_error("Cached file is longer than expected ({} B), copying aborted", model_file.size)
-                    return False
-                cache_sha256.update(data)
-                destination_file.write(data)
-        if cache_size < model_file.size:
-            reporter.log_error("Cached file is shorter ({} B) than expected ({} B)", cache_size, model_file.size)
-            return False
-        return verify_hash(reporter, cache_sha256.digest(), model_file.sha256, path)
-
-    def put(self, hash, path):
-        staging_path = None
-
-        try:
-            # A file in the cache must have the hash implied by its name. So when we upload a file,
-            # we first copy it to a temporary file and then atomically move it to the desired name.
-            # This prevents interrupted runs from corrupting the cache.
-            with path.open('rb') as src_file:
-                with tempfile.NamedTemporaryFile(dir=str(self._staging_dir), delete=False) as staging_file:
-                    staging_path = Path(staging_file.name)
-                    shutil.copyfileobj(src_file, staging_file)
-
-            hash_path = self._hash_path(hash)
-            hash_path.parent.mkdir(parents=True, exist_ok=True)
-            staging_path.replace(self._hash_path(hash))
-            staging_path = None
-        finally:
-            # If we failed to complete our temporary file or to move it into place,
-            # get rid of it.
-            if staging_path:
-                staging_path.unlink()
-
-def try_retrieve_from_cache(reporter, cache, model_file, destination):
-    try:
-        if cache.has(model_file.sha256):
-            reporter.job_context.check_interrupted()
-
-            reporter.print_section_heading('Retrieving {} from the cache', destination)
-            if not cache.get(model_file, destination, reporter):
-                reporter.print('Will retry from the original source.')
-                reporter.print()
-                return False
-            reporter.print()
-            return True
-    except Exception:
-        reporter.log_warning('Cache retrieval failed; falling back to downloading', exc_info=True)
-        reporter.print()
-
-    return False
-
-def try_update_cache(reporter, cache, hash, source):
-    try:
-        cache.put(hash, source)
-    except Exception:
-        reporter.log_warning('Failed to update the cache', exc_info=True)
-
-def try_retrieve(reporter, destination, model_file, cache, num_attempts, start_download):
-    destination.parent.mkdir(parents=True, exist_ok=True)
-
-    if try_retrieve_from_cache(reporter, cache, model_file, destination):
-        return True
-
-    reporter.print_section_heading('Downloading {}', destination)
-
-    success = False
-
-    with destination.open('w+b') as f:
-        actual_hash = try_download(reporter, f, num_attempts, start_download, model_file.size)
-
-    if actual_hash and verify_hash(reporter, actual_hash, model_file.sha256, destination):
-        try_update_cache(reporter, cache, model_file.sha256, destination)
-        success = True
-
-    reporter.print()
-    return success
-
-def download_model(reporter, args, cache, session_factory, requested_precisions, model):
-    telemetry = _common.Telemetry()
-    session = session_factory()
-
-    reporter.print_group_heading('Downloading {}', model.name)
-
-    reporter.emit_event('model_download_begin', model=model.name, num_files=len(model.files))
-
-    output = args.output_dir / model.subdirectory
-    output.mkdir(parents=True, exist_ok=True)
-
-    for model_file in model.files:
-        if len(model_file.name.parts) == 2:
-            p = model_file.name.parts[0]
-            if p in _common.KNOWN_PRECISIONS and p not in requested_precisions:
-                continue
-
-        model_file_reporter = reporter.with_event_context(model=model.name, model_file=model_file.name.as_posix())
-        model_file_reporter.emit_event('model_file_download_begin', size=model_file.size)
-
-        destination = output / model_file.name
-
-        if not try_retrieve(model_file_reporter, destination, model_file, cache, args.num_attempts,
-                functools.partial(model_file.source.start_download, session, CHUNK_SIZE)):
-            try:
-                destination.unlink()
-            except FileNotFoundError:
-                pass
-
-            model_file_reporter.emit_event('model_file_download_end', successful=False)
-            reporter.emit_event('model_download_end', model=model.name, successful=False)
-            telemetry.send_event('md', 'downloader_failed_models', model.name)
-            return False
-
-        model_file_reporter.emit_event('model_file_download_end', successful=True)
-
-    reporter.emit_event('model_download_end', model=model.name, successful=True)
-
-    if model.postprocessing:
-        reporter.emit_event('model_postprocessing_begin', model=model.name)
-
-        for postproc in model.postprocessing:
-            postproc.apply(reporter, output)
-
-        reporter.emit_event('model_postprocessing_end', model=model.name)
-
-        reporter.print()
-
-    return True
+from open_model_zoo.model_tools.download_engine.downloader import Downloader
+from open_model_zoo.model_tools.download_engine.utils import ThreadSessionFactory
 
 
 class DownloaderArgumentParser(argparse.ArgumentParser):
@@ -298,25 +42,6 @@ def positive_int_arg(value_str):
         pass
 
     raise argparse.ArgumentTypeError('must be a positive integer (got {!r})'.format(value_str))
-
-
-# There is no evidence that the requests.Session class is thread-safe,
-# so for safety, we use one Session per thread. This class ensures that
-# each thread gets its own Session.
-class ThreadSessionFactory:
-    def __init__(self, exit_stack):
-        self._lock = threading.Lock()
-        self._thread_local = threading.local()
-        self._exit_stack = exit_stack
-
-    def __call__(self):
-        try:
-            session = self._thread_local.session
-        except AttributeError:
-            with self._lock: # ExitStack might not be thread-safe either
-                session = self._exit_stack.enter_context(requests.Session())
-            self._thread_local.session = session
-        return session
 
 
 def main():
@@ -351,10 +76,11 @@ def main():
             enable_json_output=args.progress_format == 'json')
 
     reporter = make_reporter(_reporting.DirectOutputContext())
-    cache = NullCache() if args.cache_dir is None else DirCache(args.cache_dir)
 
     with _common.telemetry_session('Model Downloader', 'downloader') as telemetry:
-        models = _configuration.load_models_from_args(parser, args)
+        _configuration.check_model_selection_args(parser, args)
+        models = _configuration.load_models_from_args(args, _common.MODEL_ROOT)
+
         for mode in ['all', 'list', 'name']:
             if getattr(args, mode):
                 telemetry.send_event('md', 'downloader_selection_mode', mode)
@@ -379,15 +105,17 @@ def main():
         if unknown_precisions:
             sys.exit('Unknown precisions specified: {}.'.format(', '.join(sorted(unknown_precisions))))
 
+        downloader = Downloader(args.output_dir, args.cache_dir, args.num_attempts)
+
         with contextlib.ExitStack() as exit_stack:
             session_factory = ThreadSessionFactory(exit_stack)
             if args.jobs == 1:
-                results = [download_model(reporter, args, cache, session_factory, requested_precisions, model)
+                results = [downloader.download_model(reporter, session_factory, requested_precisions, model)
                     for model in models]
             else:
                 results = _concurrency.run_in_parallel(args.jobs,
-                    lambda context, model: download_model(
-                        make_reporter(context), args, cache, session_factory, requested_precisions, model),
+                    lambda context, model: downloader.download_model(
+                        make_reporter(context), session_factory, requested_precisions, model),
                     models)
 
         failed_models = {model.name for model, successful in zip(models, results) if not successful}
@@ -396,6 +124,7 @@ def main():
             reporter.print('FAILED:')
             for failed_model_name in failed_models:
                 reporter.print(failed_model_name)
+                telemetry.send_event('md', 'downloader_failed_models', failed_model_name)
             sys.exit(1)
 
 if __name__ == '__main__':
